starting BIM folder and refactor Door namespace
