+++ Elements/src/BIM/Door/DoorOpeningSide.cs
@@ -1,0 +1,14 @@
+namespace Elements.BIM
+{
+    public enum DoorOpeningSide
+    {
+        [System.Runtime.Serialization.EnumMember(Value = @"Undefined")]
+        Undefined,
+        [System.Runtime.Serialization.EnumMember(Value = @"Left Hand")]
+        LeftHand,
+        [System.Runtime.Serialization.EnumMember(Value = @"Right Hand")]
+        RightHand,
+        [System.Runtime.Serialization.EnumMember(Value = @"Double Door")]
+        DoubleDoor
+    }
+}
