+++ Elements/src/BIM/Door/DoorOpeningType.cs
@@ -1,0 +1,12 @@
+namespace Elements.BIM
+{
+    public enum DoorOpeningType
+    {
+        [System.Runtime.Serialization.EnumMember(Value = @"Undefined")]
+        Undefined,
+        [System.Runtime.Serialization.EnumMember(Value = @"Single Swing")]
+        SingleSwing,
+        [System.Runtime.Serialization.EnumMember(Value = @"Double Swing")]
+        DoubleSwing
+    }
+}
