+++ Elements/src/StandardWall.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Elements.Geometry;
 using Elements.Geometry.Solids;
+using Elements.BIM;
 
 namespace Elements
 {
